test: don't try to kill selenium server when not run

+++ packages/test-runner-commands/test/send-mouse/sendMousePlugin.test.ts
@@ -41,7 +41,9 @@ describe('sendMousePlugin', function test() {
     });
 
     after(() => {
-      seleniumServer.kill();
+      if (seleniumServer) {
+        seleniumServer.kill();
+      }
     });
 
     it('can send mouse on webdriver', async () => {
